refactor: moved datatypes to their own files/submodules
